Move sources into different packages
