+++ src/config/logger.js
@@ -1,5 +1,5 @@
 const log4js = require('log4js');
-const utils = require('./utils');
+const utils = require('../core/utils');
 
 const logConfigs = {
   appenders: {
