+++ src/core/config.js
